Fix for WFGP-147, Allow to package module as add-ons and layers
